+++ src/main/kotlin/com/kos/clients/raiderio/RaiderIoClient.kt
@@ -4,6 +4,7 @@ import arrow.core.Either
 import com.kos.clients.ClientError
 import com.kos.clients.domain.ExpansionSeasons
 import com.kos.clients.domain.RaiderIoCutoff
+import com.kos.clients.domain.RunDetails
 import com.kos.clients.domain.RaiderIoResponse
 import com.kos.clients.domain.RaiderioWowHeadEmbeddedResponse
 import com.kos.entities.domain.WowEntity
@@ -12,6 +13,7 @@ import com.kos.entities.domain.WowEntityRequest
 interface RaiderIoClient {
     suspend fun get(wowEntity: WowEntity): Either<ClientError, RaiderIoResponse>
     suspend fun getExpansionSeasons(expansionId: Int): Either<ClientError, ExpansionSeasons>
+    suspend fun getRunDetails(season: String, runId: String): Either<ClientError, RunDetails>
 
     suspend fun exists(wowEntityRequest: WowEntityRequest): Boolean
     suspend fun cutoff(): Either<ClientError, RaiderIoCutoff>