+++ src/main/resources/db/migration/prod/V1774828800__Alter_mythic_plus_seasons_add_slug.sql
@@ -1,0 +1,2 @@
+ALTER TABLE mythic_plus_seasons
+    ADD COLUMN slug varchar(64) NOT NULL DEFAULT '';